requirements.txt
openai==0.26.0
nltk==3.8
fpdf==1.7.2
gTTS==2.3.0
keybert==0.7.0
moviepy==1.0.3
mutagen==1.46.0
nltk==3.8
Pillow==9.3.0